# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nutrio - Indian food health scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NutritionPer100g {
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  fat?: number | null;
  /** @nullable */
  saturatedFat?: number | null;
  /** @nullable */
  carbohydrates?: number | null;
  /** @nullable */
  sugars?: number | null;
  /** @nullable */
  fiber?: number | null;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  salt?: number | null;
  /** @nullable */
  sodium?: number | null;
}

export interface Product {
  barcode: string;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** A, B, C, D, or E */
  nutriScore: string;
  /** @nullable */
  isVeg?: boolean | null;
  /** @nullable */
  isVegan?: boolean | null;
  /** @nullable */
  isSwadeshi?: boolean | null;
  /** @nullable */
  isUltraProcessed?: boolean | null;
  nutrition?: NutritionPer100g;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  servingSize?: string | null;
  /** local | openfoodfacts */
  source?: string;
}

export interface FruitVeggie {
  id: string;
  name: string;
  nameHindi: string;
  /** fruit | vegetable | leafy | root */
  category: string;
  imageEmoji?: string;
  nutrition: NutritionPer100g;
  freshnessTips?: string[];
  /** @nullable */
  season?: string | null;
  benefits?: string[];
}

export interface Ingredient {
  id: string;
  name: string;
  nameHindi: string;
  /** grain | lentil | vegetable | dairy | oil | spice | protein */
  category: string;
  nutrition: NutritionPer100g;
  defaultServingGrams?: number;
}

export type DishIngredientsItem = {
  ingredientId: string;
  grams: number;
};

export interface Dish {
  id: string;
  name: string;
  nameHindi: string;
  /** @nullable */
  description?: string | null;
  imageEmoji?: string;
  ingredients: DishIngredientsItem[];
}

export interface MealIngredientItem {
  ingredientId: string;
  grams: number;
}

export interface MealInput {
  /** @nullable */
  mealName?: string | null;
  servings?: number;
  ingredients: MealIngredientItem[];
}

export interface TrafficLight {
  /** low | medium | high */
  level: string;
  label: string;
  /** @nullable */
  value: number | null;
}

export type NutritionResultTrafficLights = {
  fat?: TrafficLight;
  sugar?: TrafficLight;
  salt?: TrafficLight;
  calories?: TrafficLight;
};

export interface NutritionResult {
  mealName: string;
  servings: number;
  totalNutrition: NutritionPer100g;
  perServingNutrition: NutritionPer100g;
  nutriScore: string;
  trafficLights: NutritionResultTrafficLights;
  tips?: string[];
}

export type SearchProductsParams = {
q: string;
limit?: number;
};

export type SearchFruitsVeggiesParams = {
q: string;
};