# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nutrio - Indian food health scanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Look up product by barcode
      description: Looks up product in local Indian DB first, then Open Food Facts API
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/search:
    get:
      operationId: searchProducts
      tags: [products]
      summary: Search products by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    NutritionPer100g:
      type: object
      properties:
        calories:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        saturatedFat:
          type: ["number", "null"]
        carbohydrates:
          type: ["number", "null"]
        sugars:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        salt:
          type: ["number", "null"]
        sodium:
          type: ["number", "null"]
    Product:
      type: object
      required:
        - barcode
        - name
        - nutriScore
      properties:
        barcode:
          type: string
        name:
          type: string
        brand:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        nutriScore:
          type: string
          description: "A, B, C, D, or E"
        isVeg:
          type: ["boolean", "null"]
        isVegan:
          type: ["boolean", "null"]
        isSwadeshi:
          type: ["boolean", "null"]
        isUltraProcessed:
          type: ["boolean", "null"]
        nutrition:
          $ref: "#/components/schemas/NutritionPer100g"
        ingredients:
          type: ["string", "null"]
        servingSize:
          type: ["string", "null"]
        source:
          type: string
          description: "local | openfoodfacts"
...