# lib/api-zod/src/generated/types/nutritionPer100g.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nutrio - Indian food health scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface NutritionPer100g {
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  fat?: number | null;
  /** @nullable */
  saturatedFat?: number | null;
  /** @nullable */
  carbohydrates?: number | null;
  /** @nullable */
  sugars?: number | null;
  /** @nullable */
  fiber?: number | null;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  salt?: number | null;
  /** @nullable */
  sodium?: number | null;
}