# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nutrio - Indian food health scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { NutritionPer100g } from './nutritionPer100g';

export interface Product {
  barcode: string;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** A, B, C, D, or E */
  nutriScore: string;
  /** @nullable */
  isVeg?: boolean | null;
  /** @nullable */
  isVegan?: boolean | null;
  /** @nullable */
  isSwadeshi?: boolean | null;
  /** @nullable */
  isUltraProcessed?: boolean | null;
  nutrition?: NutritionPer100g;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  servingSize?: string | null;
  /** local | openfoodfacts */
  source?: string;
}